Move creation of ffi object from fft_cffi to _fft_build
